Refactor component dir
move util functions to lib dir
