+++ common/models/middleware-phase.js
@@ -5,7 +5,7 @@
 'use strict';
 
 const middlewareHandler = require('../../lib/middleware-handler');
-const WorkspaceManager = require('../../component/workspace-manager.js');
+const WorkspaceManager = require('../../lib/workspace-manager.js');
 
 /**
   * Defines a `MiddlewarePhase` configuration.
