+++ datamodel/graph/edge.js
