+++ datamodel/graph/node.js
