+++ lib/tasks.js
@@ -1,14 +1,14 @@
 'use strict';
-const DataSource = require('./datamodel/datasource');
-const Facet = require('./datamodel/facet');
-const fsUtility = require('./datamodel/util/file-utility');
+const DataSource = require('../datamodel/datasource');
+const Facet = require('../datamodel/facet');
+const fsUtility = require('../lib/util/file-utility');
 const lodash = require('lodash');
-const Model = require('./datamodel/model');
-const ModelConfig = require('./datamodel/model-config');
-const ModelMethod = require('./datamodel/model-method');
-const ModelProperty = require('./datamodel/model-property');
-const MiddlewarePhase = require('./datamodel/middleware-phase');
-const PackageDefinition = require('./datamodel/package-definition');
+const Model = require('../datamodel/model');
+const ModelConfig = require('../datamodel/model-config');
+const ModelMethod = require('../datamodel/model-method');
+const ModelProperty = require('../datamodel/model-property');
+const MiddlewarePhase = require('../datamodel/middleware-phase');
+const PackageDefinition = require('../datamodel/package-definition');
 const path = require('path');
 
 /**
