+++ lib/template-registry.js
