+++ lib/util/file-utility.js
@@ -5,7 +5,7 @@
 'use strict';
 
 const async = require('async');
-const config = require('../../config');
+const config = require('../config');
 const fs = require('fs-extra');
 const glob = require('glob');
 const path = require('path');
