+++ lib/util/processor.js
