+++ lib/workspace-handler.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const fsUtility = require('../component/datamodel/util/file-utility');
+const fsUtility = require('./util/file-utility');
 const path = require('path');
 
 class WorkspaceHandler {
