+++ lib/workspace-manager.js
@@ -1,6 +1,6 @@
 'use strict';
 const config = require('./config.json');
-const Workspace = require('./workspace.js');
+const Workspace = require('./workspace');
 const templateRegistry = require('./template-registry');
 /**
  * @class WorkspaceManager
