+++ lib/workspace.js
@@ -1,13 +1,13 @@
 'use strict';
 const config = require('./config.json');
-const DataSource = require('./datamodel/datasource');
-const Graph = require('./datamodel/graph');
+const DataSource = require('../datamodel/datasource');
+const Graph = require('../datamodel/graph');
 const clone = require('lodash').clone;
-const Model = require('./datamodel/model');
-const MiddlewarePhase = require('./datamodel/middleware-phase');
+const Model = require('../datamodel/model');
+const MiddlewarePhase = require('../datamodel/middleware-phase');
 const lodash = require('lodash');
 const path = require('path');
-const Processor = require('./datamodel/util/processor');
+const Processor = require('./util/processor');
 const Tasks = require('./tasks');
 
 /**
