+++ test/acceptance/use-cases/test-scripts/find-datasource.js
@@ -6,7 +6,7 @@ const loopback = require('loopback');
 const path = require('path');
 const testSupport = require('../../../helpers/test-support');
 const util = require('util');
-const workspaceManager = require('../../../../component/workspace-manager');
+const workspaceManager = require('../../../../lib/workspace-manager');
 
 const DataSourceDefinition = app.models.DataSourceDefinition;
